refactor: dto 디렉토리 service로 이동
- 기존에 컨트롤러에 dto가 있음에 따라, 모델에서 컨트롤러에 의존하게 되어 이동
