+++ src/main/java/org/prgms/kdtspringweek1/wallet/service/WalletService.java
@@ -1,7 +1,7 @@
 package org.prgms.kdtspringweek1.wallet.service;
 
-import org.prgms.kdtspringweek1.controller.dto.customerDto.FindCustomerResponseDto;
-import org.prgms.kdtspringweek1.controller.dto.voucherDto.FindVoucherResponseDto;
+import org.prgms.kdtspringweek1.customer.service.dto.FindCustomerResponseDto;
+import org.prgms.kdtspringweek1.voucher.service.dto.FindVoucherResponseDto;
 import org.prgms.kdtspringweek1.wallet.entity.Wallet;
 import org.prgms.kdtspringweek1.wallet.repository.WalletRepository;
 import org.springframework.stereotype.Service;
