+++ src/test/java/org/prgms/kdtspringweek1/voucher/service/VoucherServiceTest.java
@@ -3,7 +3,7 @@ package org.prgms.kdtspringweek1.voucher.service;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
-import org.prgms.kdtspringweek1.controller.dto.voucherDto.FindVoucherResponseDto;
+import org.prgms.kdtspringweek1.voucher.service.dto.FindVoucherResponseDto;
 import org.prgms.kdtspringweek1.voucher.entity.FixedAmountVoucher;
 import org.prgms.kdtspringweek1.voucher.entity.PercentDiscountVoucher;
 import org.prgms.kdtspringweek1.voucher.entity.Voucher;
